+++ src/main/java/org/eclipse/yasson/internal/ConstructorPropertiesAnnotationIntrospector.java
@@ -1,10 +1,10 @@
 package org.eclipse.yasson.internal;
 
-import org.eclipse.yasson.internal.components.JsonbComponentInstanceCreator;
 import org.eclipse.yasson.internal.model.CreatorModel;
 import org.eclipse.yasson.internal.model.JsonbCreator;
 import org.eclipse.yasson.internal.properties.MessageKeys;
 import org.eclipse.yasson.internal.properties.Messages;
+import org.eclipse.yasson.spi.JsonbComponentInstanceCreator;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Executable;
