+++ src/main/resources/yasson-messages.properties
@@ -80,4 +80,3 @@ datatypeFactoryCreationFailed=An error occurred while DatatypeFactory creation.
 multipleConstructorPropertiesCreators=More than one constructor annotated with @ConstructorProperties declared in class {0}.
 annotationNotAvailable=Annotation {0} is not visible in modules or classpath. Annotation will be ignored.
 missingValuePropertyInAnnotation=Missing value property in Annotation {0}. Annotation will be ignored.
-usingServiceProviderComponentInstanceCreator=Component instance creator: {0}
